Put Identifier in its own file; namespacing
