+++ catalogue_pipeline/id_minter/src/main/scala/uk/ac/wellcome/platform/idminter/models/IdentifiersTable.scala
@@ -1,0 +1,22 @@
+package uk.ac.wellcome.platform.idminter.models
+
+import com.google.inject.Inject
+import com.twitter.inject.annotations.Flag
+import scalikejdbc._
+
+class IdentifiersTable @Inject()(
+  @Flag("aws.rds.identifiers.database") database: String,
+  @Flag("aws.rds.identifiers.table") table: String)
+    extends SQLSyntaxSupport[Identifier] {
+  override val schemaName = Some(database)
+  override val tableName = table
+  override val useSnakeCaseColumnName = false
+  override val columns = Seq(
+    "CanonicalID",
+    "ontologyType",
+    "MiroID",
+    "CalmAltRefNo"
+  )
+
+  val i = this.syntax("i")
+}
